refactor(comp:divider): rebuild with tsx
